requirements.txt
numpy>=1.21.0
gymnasium>=0.29.0
PyOpenGL>=3.1.6
Pygame>=2.5.0
# Optional dependencies for RL algorithm examples
# stable-baselines3>=1.5.0  # Uncomment if using Stable Baselines 3
# ray[rllib]>=2.0.0  # Uncomment if using RLlib 